# perf: 优化开启多行文本时的布局性能