--- a/packages/s2-core/src/cell/col-cell.ts
+++ b/packages/s2-core/src/cell/col-cell.ts
@@ -82,7 +82,7 @@ export class ColCell extends HeaderCell<ColHeaderConfig> {
     return super.getFormattedFieldValue();
   }
 
-  protected getMaxTextWidth(): number {
+  public getMaxTextWidth(): number {
     const { width } = this.getBBoxByType(CellClipBox.CONTENT_BOX);
 
     return width - this.getActionAndConditionIconWidth();