--- a/packages/s2-core/src/common/interface/basic.ts
+++ b/packages/s2-core/src/common/interface/basic.ts
@@ -440,6 +440,7 @@ export interface ViewMeta {
   value?: string | number;
   query?: Query;
   isLeaf?: boolean;
+  shallowRender?: boolean;
 
   [key: string]: unknown;
 }