--- a/packages/s2-core/src/facet/frozen-facet.ts
+++ b/packages/s2-core/src/facet/frozen-facet.ts
@@ -303,7 +303,10 @@ export abstract class FrozenFacet extends BaseFacet {
 
     if (viewMeta) {
       viewMeta.isFrozenCorner = true;
-      const cell = this.spreadsheet.options.dataCell?.(viewMeta)!;
+      const cell = this.spreadsheet.options.dataCell?.(
+        viewMeta,
+        this.spreadsheet,
+      )!;
 
       group.appendChild(cell);
     }