--- a/packages/s2-core/src/interaction/brush-selection/data-cell-brush-selection.ts
+++ b/packages/s2-core/src/interaction/brush-selection/data-cell-brush-selection.ts
@@ -136,7 +136,7 @@ export class DataCellBrushSelection extends BaseBrushSelection {
         meta.colIndex,
       )!;
 
-      return this.spreadsheet.options.dataCell!(viewMeta);
+      return this.spreadsheet.options.dataCell!(viewMeta, this.spreadsheet);
     });
   }