… packages/s2-core/__tests__/benchmark/pivot-sheet-spec.ts
… packages/s2-core/__tests__/benchmark/table-sheet-spec.ts
… packages/s2-core/__tests__/spreadsheet/__snapshots__/multi-line-text-spec.ts.snap
… packages/s2-core/__tests__/spreadsheet/multi-line-text-spec.ts
… packages/s2-core/package.json
… packages/s2-core/src/utils/interaction/merge-cell.ts
… packages/s2-react/package.json
… packages/s2-react/scripts/test-live.mjs
… packages/s2-shared/package.json
… packages/s2-vue/package.json
… pnpm-lock.yaml