### 👀 PR includes

<!-- Add completed items in this PR, and change [ ] to [x]. -->
<!-- 这个 PR 属于什么类型，请选中对应类型 [ ] to [x]. -->

🎨 Enhance

- [ ] Code style optimization
- [ ] Refactoring
- [ ] Change the UI
- [x] Improve the performance
- [ ] Type optimization

### 📝 Description

<!-- What is this PR change point? What is the background? What problems have been solved? -->
<!-- 这个 PR 改动点是什么？背景是什么？解决了什么问题？-->

在 https://github.com/antvis/S2/pull/2728 中优化了没有开启多行文本时的性能, 这个 PR 优化开启多行文本时的布局性能  (完成 https://github.com/antvis/S2/pull/2728#issuecomment-2117158843 中的 TODO)

1. 使用同一个 class 的示例, 动态更新 meta 的方式, 避免频繁创建实例 GC 的性能问题.
2. 对文本长度相同, 并且单元格宽度一致的情况, 增加缓存, 大幅降低多行文本高度测量的性能损耗.
3. 减少明细表频繁 `setOptions` 触发 merge 的性能损耗.
4. 明细表适配 `shallowRender` 逻辑, 减少 `initCell` 的性能损耗.

测试数据样本 (2w 条明细数据):

![image](https://github.com/antvis/S2/assets/21015895/0f24d77c-8137-4fa6-97f9-57e880562836)


### 🖼️ Screenshot

<!-- Comparison of screenshots before and after changes, it is best to be GIF -->
<!-- 改动前后的截图对比，最好是 GIF -->

| Before | After |
| ------ | ----- |
| 透视表 (9.8s -> 1.8s) ![image](https://github.com/antvis/S2/assets/21015895/2087bfdc-f54b-4477-b8ac-e215c09199f8) | ![image](https://github.com/antvis/S2/assets/21015895/0e598125-886d-41ef-af6a-c6271a6ae5d0) |
| 明细表 (16s -> 0.35s (有缓存) -> 5.7s (无缓存)) ![image](https://github.com/antvis/S2/assets/21015895/a24081bd-200f-42a2-8f57-a9ae569c7a1b) | ![image](https://github.com/antvis/S2/assets/21015895/7ed90b76-5897-4b39-8acf-e0b2f401f236) ![image](https://github.com/antvis/S2/assets/21015895/50bc89ab-a335-4917-acd8-143e75d70fe5) |



### 🔗 Related issue link

<!-- If there is a related Issue/PR link -->
<!-- 如果有相关的 Issue/PR 链接，请关联上 -->

<!-- close #0 -->
<!-- ref #0 -->
<!-- fix #0 -->

### 🔍 Self-Check before the merge

<!-- Please add test case, docs, and demos -->
<!-- 吾日三省吾身，有添加单元测试吗？有完善文档吗？有增加文档示例吗？-->

- [ ] Add or update relevant docs.
- [ ] Add or update relevant demos.
- [ ] Add or update test case.
- [ ] Add or update relevant TypeScript definitions.