@lijinke666 2024-05-24
:tada: This PR is included in version @antv/s2-react-v2.0.0-next.21 :tada:

The release is available on:
- [npm package (@next dist-tag)](https://www.npmjs.com/package/@antv/s2-react/v/2.0.0-next.21)
- [GitHub release](https://github.com/antvis/S2/releases/tag/%40antv/s2-react-v2.0.0-next.21)

Your **[semantic-release](https://github.com/semantic-release/semantic-release)** bot :package::rocket: